feat: added code for handling phone account deletion, pending TODOs for frontend

+++ frontend/screens/Profile.js
@@ -21,6 +21,7 @@ import screenStyles from '../../styles/screenStyles.js'
 import Settings from '../modals/ProfileSettings.js'
 import socket from '../apis/socket.js'
 import TabBar from '../Nav.js'
+import auth from '@react-native-firebase/auth'
 
 class UserProfileView extends Component {
   constructor(props) {
@@ -44,6 +45,8 @@ class UserProfileView extends Component {
       numFriends: 0,
       imageData: null,
       disabled: false,
+      // phone auth
+      verificationId: null,
     }
   }
   // getting current user's info
@@ -96,12 +99,49 @@ class UserProfileView extends Component {
     }
   }
 
+  ////////////////////////////////////////////////////////////////////////////////
+  ///////////////////////////////////PHONE AUTH///////////////////////////////////
+  ////////////////////////////////////////////////////////////////////////////////
+
+  // TODO: Create code input modal that for 6 digit code that executes the "verifyCode" function (below) upon submitting the code
+  // More TODOs in functions below
+
   async handleDelete() {
     if (!this.state.disabled) {
       this.setState({ disabled: true })
-      loginService
+      // Check if phone provider - if so, validate phone num
+      if (auth().currentUser.providerData[0].providerId === 'phone')
+      {
+        auth()
+        .verifyPhoneNumber(auth().currentUser.phoneNumber)
+        .on('state_changed', (phoneAuthSnapshot) => {
+          console.log('State: ', phoneAuthSnapshot.state)
+          console.log(phoneAuthSnapshot)
+          switch (phoneAuthSnapshot.state) {
+            // Ask for code input
+            case auth.PhoneAuthState.CODE_SENT:
+              // Store verification id in state
+              this.setState({verificationId: phoneAuthSnapshot.verificationId})
+              this.verifyCode('111111')
+// TODO: Display verification code input modal (6 digits)
+              break;
+            // Auto verified on android - proceed to delete account
+            case auth.PhoneAuthState.AUTO_VERIFIED:
+              this.setState({verificationId: phoneAuthSnapshot.verificationId})
+              this.verifyCode(phoneAuthSnapshot.code)
+              break;
+            // Display error alert
+            case auth.PhoneAuthState.ERROR:
+// TODO: Code could not be sent, display an error
+              this.props.showError()
+              break;
+          }
+        })
+      }
+      // Not phone provider - delete the account
+      else {
+        loginService
         .deleteUser()
-        // TODO: Disastrous phone auth code...
         .then(() => {
           // close settings and navigate to Login
           this.setState({ visible: false })
@@ -110,9 +150,42 @@ class UserProfileView extends Component {
         .catch(() => {
           this.props.hideError()
         })
+      }
       this.setState({ disabled: false })
     }
   }
+
+  // IMPORTANT: Code MUST be passed in as a string
+  // NOTE: Currently only logs success. To delete, uncomment lines under log.
+  async verifyCode(code) {
+    console.log("Verifying code ", code)
+    const credential = auth.PhoneAuthProvider.credential(this.state.verificationId, code)
+    auth().currentUser.reauthenticateWithCredential(credential)
+      // If reauthentication succeeds, delete account using credential
+      .then(() => {
+        console.log("success")
+        // loginService
+        // .deleteUserWithCredential(credential)
+        // .then(() => {
+        //   // close settings and navigate to Login
+        //   this.setState({ visible: false })
+        //   this.props.navigation.replace('Login')
+        // })
+        // .catch(() => {
+        //   this.props.hideError()
+        // })
+      })
+      .catch((error) => {
+        console.log("Code verification failed")
+        console.log(error)
+// TODO: Code could not be verified, disply an error
+        this.props.showError()
+      })
+  }
+
+  ////////////////////////////////////////////////////////////////////////////////
+  ////////////////////////////////////////////////////////////////////////////////
+  ////////////////////////////////////////////////////////////////////////////////
 
   // close alert for taken username
   closeTaken() {
